разнес ui по папкам (поменял структуру)
