+++ app/src/main/java/com/example/moviesearch/view/MainActivity.kt
@@ -5,6 +5,9 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import com.example.moviesearch.R
 import com.example.moviesearch.databinding.MainActivityBinding
+import com.example.moviesearch.view.main.ListFragment
+import com.example.moviesearch.view.favorites.FavoritesMoviesFragment
+import com.example.moviesearch.view.settings.SettingsFragment
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
 class MainActivity : AppCompatActivity() {
