+++ app/src/main/java/com/example/moviesearch/view/details/DetailsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.moviesearch.view
+package com.example.moviesearch.view.details
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,24 +6,25 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.example.moviesearch.databinding.FragmentSettingsBinding
+import com.example.moviesearch.databinding.FragmentDetailsBinding
 import com.example.moviesearch.viewmodel.MainViewModel
 
-class SettingsFragment : Fragment() {
+
+class DetailsFragment : Fragment() {
 
     companion object {
-        fun newInstance() = SettingsFragment()
+        fun newInstance() = DetailsFragment()
     }
 
     private lateinit var viewModel: MainViewModel
-    var _binding: FragmentSettingsBinding? = null
+    var _binding: FragmentDetailsBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
     ): View {
 
-        _binding = FragmentSettingsBinding.inflate(inflater, container, false)
+        _binding = FragmentDetailsBinding.inflate(inflater, container, false)
         return binding.root
     }
