+++ app/src/main/java/com/example/moviesearch/view/main/ListFragment.kt
@@ -1,4 +1,4 @@
-package com.example.moviesearch.view
+package com.example.moviesearch.view.main
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,26 +6,27 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.example.moviesearch.databinding.FragmentDetailsBinding
+import com.example.moviesearch.databinding.FragmentListBinding
 import com.example.moviesearch.viewmodel.MainViewModel
 
-
-class DetailsFragment : Fragment() {
+class ListFragment : Fragment() {
 
     companion object {
-        fun newInstance() = DetailsFragment()
+        fun newInstance() = ListFragment()
     }
 
     private lateinit var viewModel: MainViewModel
-    var _binding: FragmentDetailsBinding? = null
+    private var _binding: FragmentListBinding? = null
+
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
     ): View {
 
-        _binding = FragmentDetailsBinding.inflate(inflater, container, false)
-        return binding.root
+        _binding = FragmentListBinding.inflate(inflater, container, false)
+        val view = binding.root
+        return view
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -38,4 +39,5 @@ class DetailsFragment : Fragment() {
         super.onDestroyView()
         _binding = null
     }
+
 }
