+++ app/src/main/java/com/example/moviesearch/view/settings/SettingsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.moviesearch.view
+package com.example.moviesearch.view.settings
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,27 +6,25 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.example.moviesearch.databinding.FragmentListBinding
+import com.example.moviesearch.databinding.FragmentSettingsBinding
 import com.example.moviesearch.viewmodel.MainViewModel
 
-class ListFragment : Fragment() {
+class SettingsFragment : Fragment() {
 
     companion object {
-        fun newInstance() = ListFragment()
+        fun newInstance() = SettingsFragment()
     }
 
     private lateinit var viewModel: MainViewModel
-    private var _binding: FragmentListBinding? = null
-
+    var _binding: FragmentSettingsBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
     ): View {
 
-        _binding = FragmentListBinding.inflate(inflater, container, false)
-        val view = binding.root
-        return view
+        _binding = FragmentSettingsBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -39,5 +37,4 @@ class ListFragment : Fragment() {
         super.onDestroyView()
         _binding = null
     }
-
 }
