+++ app/src/main/res/layout/fragment_favorites_movies.xml
@@ -4,7 +4,7 @@
     android:id="@+id/fragment_favorite_movies"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.FavoritesMoviesFragment">
+    tools:context=".view.favorites.FavoritesMoviesFragment">
 
     <TextView
         android:id="@+id/text"
