+++ app/src/main/res/layout/fragment_list.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.ListFragment">
+    tools:context=".view.main.ListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/listFragmentRecyclerView"
