+++ app/src/main/res/layout/fragment_settings.xml
@@ -4,7 +4,7 @@
     android:id="@+id/fragment_settings"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.SettingsFragment">
+    tools:context=".view.settings.SettingsFragment">
 
     <TextView
         android:layout_width="match_parent"
